Rename parse_choice.py -> choices.py.
